moved QuestionComponentFactory to QuestionMaker
